requirements.txt
PyQt6>=6.6.0
httpx>=0.27.0
pyobjc-framework-Cocoa>=12.0
pyobjc-framework-Quartz>=12.0
